feat: add a functional test bot to the library
This functional test can be used to quickly check whether the library
works, because it spawns a sample bot that can be used to quickly test
integration against the library. Provide the required config as
environment variables:

* BOT_TOKEN: the token to login
* ROLE: the role to grant the member when enters the voice channel
* CHANNEL: the voice channel to monitor - the bot must be able to see it

Make the bot connect into a guild that has a voice channel properly
configured, and it should grant the user the role when joins the
channel, and remove from the role when leaves.