fix: do not trigger actions on mute/deafen
Voice states are also triggered when the user mutes, unmutes, deafs or
undeafs. In this case, both old and new voice states will have the same
channel ID.

It was reported that when an user mutes or unmutes, it still causes the
roles to momentarilly change, which may be noticed because the color of
the user flashes.

This commit will only trigger a role change when the channel IDs do not
match, ignoring any case of matching channel IDs (which would mean that
the member has been muted or deafened).

Related: https://github.com/makigas/clank/issues/445