refactor!: manager no longer a default export
This is a breaking change that will probably require you to change the
way discordjs-voicerole is imported in TypeScript projects. For
JavaScript projects, it addresses an important bug with the way to
import this project, since default exports in TypeScript do not play
nice in JavaScript.

The VoiceRoleManager is no longer a default export and instead it is a
normal export now, which means that the proper way to import this now in
your project is:

For TypeScript / ES Module projects:

    import { VoiceRoleManager } from "discordjs-voicerole";

For CommonJS projects:

    const { VoiceRoleManager } = require("discordjs-voicerole");